+++ packages/terra-application/src/terra-dev-site/test/modal-manager/DisclosureComponent.jsx
@@ -11,12 +11,11 @@ import {
   DisclosureManagerHeaderAdapter,
   closeMostRecentDisclosure,
   getActiveDisclosureCount,
-} from 'terra-application/lib/disclosure-manager';
-import NavigationPrompt from 'terra-application/lib/navigation-prompt';
-import ApplicationLoadingOverlay from 'terra-application/lib/application-loading-overlay';
-import ApplicationStatusOverlay from 'terra-application/lib/application-status-overlay';
-
+} from '../../../disclosure-manager';
+import NavigationPrompt from '../../../navigation-prompt';
+import ApplicationLoadingOverlay from '../../../application-loading-overlay';
 import NotificationBannerExample from '../notification-banner/NotificationBannerExample';
+import ApplicationStatusOverlay from '../../../application-status-overlay';
 
 import styles from './DisclosureComponentCommon.test.module.scss';