+++ packages/terra-application/src/terra-dev-site/test/workspace/Tab5.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import {
-  WorkspaceContent, WorkspaceContentActivityOverlay, WorkspaceContentStatusOverlay, WorkspaceContentStatusOverlayButton,
-} from 'terra-application/lib/workspace';
+  WorkspaceContent,
+} from '../../../workspace';
 
 const Tab5 = () => {
   const [isLoading, setIsLoading] = React.useState();
 
   return (
     <WorkspaceContent
-      activityOverlay={isLoading ? <WorkspaceContentActivityOverlay variant="loading" /> : undefined}
+      activityOverlay={isLoading ? <WorkspaceContent.ActivityOverlay variant="loading" /> : undefined}
       statusOverlay={(
-        <WorkspaceContentStatusOverlay variant="error" message="This is an error message.">
-          <WorkspaceContentStatusOverlayButton
+        <WorkspaceContent.StatusOverlay variant="error" message="This is an error message.">
+          <WorkspaceContent.StatusOverlay.Button
             text="Retry"
             onClick={() => {
               setIsLoading(true);
@@ -21,12 +21,10 @@ const Tab5 = () => {
               }, 5000);
             }}
           />
-        </WorkspaceContentStatusOverlay>
+        </WorkspaceContent.StatusOverlay>
       )}
     />
   );
 };
-
-Tab5.titleKey = 'derp';
 
 export default Tab5;