+++ packages/terra-application/src/terra-dev-site/test/workspace/Workspace.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Workspace, { WorkspaceItem } from 'terra-application/lib/workspace';
-import ActiveMainPageContext from 'terra-application/lib/application-container/private/active-main-page/ActiveMainPageContext';
+import Workspace, { WorkspaceItem } from '../../../workspace';
+import ActiveMainPageContext from '../../../application-container/private/active-main-page/ActiveMainPageContext';
 import Tab1 from './Tab1';
 import Tab2 from './Tab2';
 import Tab3 from './Tab3';